Move acquisition into a separate function
We fixed offline because if forceUpdate is true with a caller but we are offline we still dont want to try to update because we can't.

This moves the acquire local to a separate function because we will want to call it to 'update' in the install step of the update. We dont want the overhead of calling into the vscode command api, plus this will allow for further customization such as silencing the acquisition message.